requirements.txt
pymupdf==1.22.5 
pytesseract==0.3.10  
Pillow==9.5.0  
camelot-py[cv]==0.10.1 
pdfminer.six==20221105 